Move util function to utils module
