+++ js/screens/rules.js
@@ -1,7 +1,5 @@
 import {addListenersToBackButton} from './common.js';
-import {addListenersWhenPageIsReady} from './../utils.js';
-import getElementFromTemplate from './../get-element.js';
-import switchScreen from './../switch-screen.js';
+import {addListenersWhenPageIsReady, getElementFromTemplate, switchScreen} from './../utils.js';
 
 const html = `<header class="header">
     <div class="header__back">
