+++ js/utils.js
@@ -18,4 +18,31 @@ const addListenersWhenPageIsReady = (cb, selector) => {
   });
 };
 
-export {addListenersWhenPageIsReady};
+/**
+ * Returns new DOM element created by passed string
+ *
+ * @param {String} elementString
+ * @return {Element} div
+ * **/
+const getElementFromTemplate = (elementString) => {
+  let fragment = document.createElement(`template`);
+  fragment.innerHTML = elementString.trim();
+  return fragment.content;
+};
+
+/**
+ * Change index of screen to the new one and render screen with new index
+ *
+ * @param {Element} newScreenElement
+ * **/
+const switchScreen = (newScreenElement) => {
+  const mainElement = document.querySelector(`main.central`);
+  mainElement.innerHTML = ``;
+  mainElement.appendChild(newScreenElement);
+};
+
+export {
+  addListenersWhenPageIsReady,
+  getElementFromTemplate,
+  switchScreen
+};
